+++ backend/homeggu-goods/src/main/java/com/homeggu/domain/salesBoard/SalesBoardDTO.java
@@ -16,8 +16,8 @@ import java.util.List;
 @Builder
 public class SalesBoardDTO {
 
-    private Integer salesBoardId;
-    private Integer userId;
+    private Long salesBoardId;
+    private Long userId;
     private String title;
     private String content;
     private Category category;
